📦 Move unconditional dependencies into extras
The extra names in the core packaging metadata are allowed to contain
dashes but not underscores [[1]]. Packaging tooling is supposed to
normalize them if met in the non-metadata configuration but apparently
tox has a bug connected to this, which prevents is from discovering
dependencies provided by extras that contain underscores [[2]].
So as a workaround, this patch uses hyphens in the names of extras,
avoiding any underscores.

Making the direct dependencies optional, means that they are not going
to be picked up by default. To enable pip-tools to include them into
the lock files, this change also updates its configuration file to
instruct it to include dependencies from all the extras present by
always using `--all-extras`. Likewise, to enable tox to install those
dependencies, they are all listed in its configuration file too.

Co-Authored-By: Sviatoslav Sydorenko <webknjaz@redhat.com>

[1]: https://packaging.python.org/en/latest/specifications/core-metadata/#provides-extra-multiple-use
[2]: https://github.com/tox-dev/tox/issues/3433#issuecomment-2452769057